Fix destruction of suspended generators in suspended fibers during shutdown
The generator destructor is a no-op when the generator is running in a fiber,
because unwinding the fiber will resume the generator. Normally the destructor
is never called in this case, but this can happen during shutdown.

We detect that a generator is running in a fiber with the
ZEND_GENERATOR_IN_FIBER flag.

This change fixes two cases in which this mechanism was broken:

 - The flag was not added when resuming a 'yield from $nonGenerator', as this is
   handled in a separate code path
 - When a generator that is running in a fiber has multiple children (aka
   multiple generators yielding from it), all of them could be considered to
   also run in a fiber (only one actually is), and could leak if not destroyed
   before shutdown.