Fix bug #51558: shared readline build fails
The 'rl_pending_input' is a variable in Readline library and checking it
with PHP_CHECK_LIBRARY wouldn't find it on some systems. This should fix
the build on systems where this caused issues, such as AIX.

Which works on most systems but not on AIX mentioned in the bug as it
exports variables and functions differently whereas the linker couldn't
resolve the variable as a function.

The library check:

```c
char rl_pending_input ();
int main (void) {
    return rl_pending_input ();
}
```

The declaration check:

```c

int main (void) {
    (void) rl_pending_input;
    (void) rl_pending_input;

    ;
    return 0;
}
```

Closes https://bugs.php.net/51558