Autotools: Enhance phpdbg readline enabling
PHP_READLINE, PHP_LIBEDIT, and PHP_READLINE_LIBS variables are not
available in sapi before the extensions arguments get processed by the
configure script. But the Autoconf raw default variables with_readline
and with_libedit are. This checks if readline support can be enabled in
phpdbg

This is partial fix towards the more proper GH-13184 or similar in the
future to make phpdbg readline integration easier to use.

Previous check silently enabled the phpdbg readline integration when
readline is built as shared and failed on the make step:

    undefined reference to 'readline'
    undefined reference to 'add_history'

These cases enable it:
    ./configure --enable-phpdbg-readline --with-libedit
    ./configure --enable-phpdbg-readline --with-readline
These cases don't enable it:
    ./configure --enable-phpdbg-readline
    ./configure --enable-phpdbg-readline --with-libedit=shared
    ./configure --enable-phpdbg-readline --with-readline=shared