opcache: Register constants for opcache.opt_debug_level and opcache.jit_debug
Register constants for the integer values which can be used for
opcache.opt_debug_level and opcache.jit_debug.

Utility is limited since, contrary to a comment in php.ini-development,
extension constants apparently cannot be used in php.ini after the
zend_extension= line. But that can be fixed separately. With this change
it is possible to do

ini_set('opcache.jit_debug', OPCACHE_JIT_DEBUG_ASM);

And that's something at least. Perhaps more importantly, registering
these constants gives us a place to document their values in the PHP
manual -- "read the source" is unsatisfying.

I did ZEND -> OPCACHE to reflect the current organisation of the code. I
think ZEND_ in this case refers to the Zend Accelerator rather than the
Zend Engine, which is confusing.