Drop unused, but confusing macro definition
`PHP7DLLTS_EXPORTS` is not used throughout php-src, and it's unlikely
that it is used by any extension, especially since it still refers to
PHP 7.  The symbol had been introduced in 2014[1], but is not even used
in latest PHP-7.4; probably it was just supposed to be used, or part of
phpng.

[1] <https://github.com/php/php-src/commit/fc734cf95da914aa0f6eac0b8ceeb1c1d9fcdf69>