Provide function for resolving a single address
A common idiom is only caring about the first IP address, for i.e.
resolving localhost -> 127.0.0.1/::1. Iterating through and freeing the
whole list in this case is tedious.

In this case, we provide a function that takes a pointer to a sockaddr
storage (so it can be stack allocated), as well as information like
address family (for i.e. only IPv4 addresses). This should hopefully be
only as boilerplate as gethostbyname was.