Fix ZEND_EXT_API wrt accessibility of OPcache symbols on Windows
When building a DLL which exports symbols these need to be explicitly
exported like on other platforms.  When building a DLL which accesses
these symbols, they should be imported for performance reasons if the
symbols are referring to functions, and *have* *to* be imported if the
symbols are referring to data.[1]

So we do the common dllexport/dllimport "dance" depending on a macro
which is only set when building OPcache.

[1] <https://learn.microsoft.com/en-us/cpp/build/importing-into-an-application-using-declspec-dllimport?view=msvc-170>