Add labels to PRs with potential ABI breaks
ABI breaks are not supposed to happen after feature freeze, i.e. when
the PHP API numbers have been bumped.  To make it easier to notice
inadvertent ABI breaks, we automatically add an "ABI break" label to
all PRs which modify public (aka. installed) header files.  Some of
these modifications do not constitute an ABI break (e.g. adding a
comment to a header file), but we rely on natural intelligence to sort
that out.  That means these labels should be removed manually, if they
are not appropriate, but if they are, the PR should not be merged into
any stable branch.  For the master branch, where ABI breaks are
permissible, the labels should still be removed if irrelevant, but kept
when the PR is merged.

Since tests are futile[1], we leave that to further (test) PRs.

[1] <https://github.com/php/php-src/pull/15682#issuecomment-2323060586>