Refactor
We extract `report_getaddrinfo_failure()`, and refactor that further.
While this code is now a tad bit slower, that shouldn't matter since it
is error handling.

We also cheat a bit, since adding a proper format specifier for
`php_error_docref()` is not a proper refactoring, but more like a bug
fix.