Autotools: Add PHP_EXTRA_CFLAGS environment variable
This adds a new configure-phase environment variable PHP_EXTRA_CFLAGS
which are appended to the CFLAGS in the Makefile rules. It sets the
already present undocumented EXTRA_CFLAGS Makefile variable to not
introduce issues with EXTRA_CFLAGS usage in the existing PECL
extensions.

This enables setting configure-and-build-phase CFLAGS and build-phase
PHP_EXTRA_CFLAGS separately to not affect configure checks:

    ./configure CFLAGS=... PHP_EXTRA_CFLAGS=...