allow json_encode() on most anonymous classes
anonymous classes are market as ZEND_ACC_NOT_SERIALIZABLE because
unserisalizing them for later use is hard to impossible. However, in the
case jon json_encode(), as long as the parent class is serializable,
there's very little reason to forbidding to json_encode() anonymous
classes, indeed, that's probably very useful in case of userland parent
classes.