Autotools: Include only php_config.h in libmbfl
This is a sync with the include style used as of PHP-8.4 to be able to
do simultaneous in-source and out-of-source builds.

When building with phpize there isn't any symbol used from the generated
extensions's config.h but the php_config.h is needed for some
preprocessor macros to be resolved properly.